:rocket: Main feature 01

+++ client/src/components/addPlace/addImages/ImagesList.jsx
@@ -11,12 +11,13 @@ import deleteFile from '../../../firebase/deleteFile';
 
 const ImagesList = () => {
   const {
-    state: { images, currentUser },
+    state: { images, currentUser, updatedPlace },
     dispatch,
   } = useValue();
 
   const handleDelete = async (image) => {
     dispatch({ type: 'DELETE_IMAGE', payload: image });
+    if (updatedPlace) return dispatch({ type: 'UPDATE_DELETED_IMAGES', payload: [image] });
     const imageName = image?.split(`${currentUser?.id}%2F`)[1]?.split('?')[0];
     try {
       await deleteFile(`rooms/${currentUser?.id}/${imageName}`);
